Added Creator entities and Changed editorentity to editor, Editor entities need to be updated, mapping not done
