+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/dto/editordto/EditorProjectDTO.java
@@ -1,6 +1,6 @@
 package com.katziio.collabwithkatz.dto.editordto;
 
-import com.katziio.collabwithkatz.entity.editorentity.EditorProject;
+import com.katziio.collabwithkatz.entity.editor.EditorProject;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
