+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/entity/creator/Review.java
@@ -1,4 +1,4 @@
-package com.katziio.collabwithkatz.entity.editorentity;
+package com.katziio.collabwithkatz.entity.creator;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "editor_communication_language")
-public class EditorCommunicationLanguage {
+@Table(name = "review")
+public class Review {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private String language;
-    private int languageProficiency;
+    private String reviewDescription;
+
 }
