+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/entity/creator/Upvote.java
@@ -1,0 +1,18 @@
+package com.katziio.collabwithkatz.entity.creator;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "upvote")
+public class Upvote {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private Boolean isVoted;
+}
