+++ Backend/collabwithkatz/src/main/java/com/katziio/collabwithkatz/entity/editor/EditorProject.java
@@ -1,0 +1,19 @@
+package com.katziio.collabwithkatz.entity.editor;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "editor_project")
+public class EditorProject {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private String projectName;
+    private boolean projectStatus;
+}
